[autotest] RDB prefers to choose a host with the same cros-version
Once we share pool between, say canary and pfq, it is possible that
two suites can exchange duts which would lead to unnecessary cost
of dut provisioning back and forth between two builds.

This can be solved by making host-scheduler always prefer a host
with the same cros-version as the job if such a host is available.

The idea is generalized to allow the request to specify a list
of preferred deps, so that it won't be limited to only
cros-version.

DEPLOY=host_scheduler
BUG=chromium:452752
TEST=Test as follows:
- Create 10 fake duts: fake1, fake2...fake10
- Randomly select 3 and assign them with a specific cros-version
- Run a dummy suite with the cros-version
- Observe the 3 random duts are selected.
  Log the sorted list of hosts, observe the 3 hosts are ranked
  in the front.
- unittest and integration test.

Change-Id: I0d5def604523173cf59b60d802ca981b6d68bb5f
Reviewed-on: https://chromium-review.googlesource.com/243750
Reviewed-by: Mungyung Ryu <mkryu@google.com>
Commit-Queue: Fang Deng <fdeng@chromium.org>
Tested-by: Fang Deng <fdeng@chromium.org>