Trying some workarounds for exceptions that can only reproduce in the lab.
As described in issue 452683, we are getting
DevtoolsTargetCrashException exceptions that can only be reproduced in
the lab.  I added some code to try to detect when the browser dies,
wait until it recovers and continue. Also modified the code to dump
some more information.

BUG=chromium:452683
TEST=Tested locally.

Change-Id: I6814ccc98ba638a3f6056f0149dd8936ab25aaf2
Reviewed-on: https://chromium-review.googlesource.com/244113
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Luis Lozano <llozano@chromium.org>