Intepret video.currentTime as a float not as in int.
video_VideoSanity tests have been failing intermittently on skate.

The test was set up to get updated currentTime from the video and
compare it to what it received as currentTime one second ago.

However these values were casted to ints, so prevCurrentTime = 1.01
and currentTime = 1.99 would be casted to one and the test will think
currentTime is not moving. The test will fail in that case.

I suspect is happening on skate because somehow things are slow there.

Additional changes are to only call currentTime once.

BUG=chromium:450335
TEST=Ran the test on a skate.

Change-Id: If632538f2e5391f91ed9814f99b8644d2824e342
Reviewed-on: https://chromium-review.googlesource.com/245160
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Commit-Queue: Mussa Kiroga <mussa@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>