Chameleon: Raise a TestWarn when I2C bus gets wedged
We saw an issue that the hot plug noise may cause the DDC bus to
get wedged. We modified the test to replug the cable when that
happens. As it is a command behavior when an user sees a black
screen. But we should treat it as a TestWarn instead of a TestPass.

BUG=chromium:447510
TEST=Ran the test multiple time and checked a TestWarn result.

Change-Id: I04c0e61692857bffa8f28d18ff71f587faded87d
Reviewed-on: https://chromium-review.googlesource.com/246078
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Hung-ying Tyan <tyanh@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>