Map AFE boards to servo boards in infra_ServoDiagnosis
We are growing a number of variations on boards that are irrelevent from
the servo perspective.  Tolerate variations on board names coming from
the AFE for the purposed of servo validation.

Concretely, *_freon boards should map to their basic versions, and a
gizmo is just a panther.

BUG=None
TEST=unittest

Change-Id: Ifc9257df9132a8c2cc9cb387e5cf00ae36d5f169
Reviewed-on: https://chromium-review.googlesource.com/246373
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>