Set video_VideoSanity to run per_build.
video_VideoSanity was reverted first from bvt-cq as it was blocking PFQ and we
did not fully understand why the test failed intermittently.

We renabled the test but the suite was set to 'video'. We need to get more runs
so setting suite to be bvt-perbuild.

BUG=chromium:448208, chromium:454874, chromium:450335
TEST=None

Change-Id: I765be4fea001104e32db1f7a0135ce5722c066d5
Reviewed-on: https://chromium-review.googlesource.com/247158
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: Mussa Kiroga <mussa@chromium.org>
Commit-Queue: Mussa Kiroga <mussa@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>