touch_WakeSource autotest: handle devices with touchscreen but no touchpad
Some devices currently fail this test because their touchscreens are
supposed to be a wake source.  Change the logic of the test to pass for all
devices that have a touchscreen as a wakesource but no touchpad present.

BUG=None
TEST=ran it on devices with touchscreen but no touchpad and devices with both
touchscreen and touchpad.
DEPLOY=test_importer

Change-Id: I1a0a59336f92fd494642da4f9e736cb59911a355
Reviewed-on: https://chromium-review.googlesource.com/248390
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Commit-Queue: Katherine Threlkeld <kathrelkeld@chromium.org>
Tested-by: Katherine Threlkeld <kathrelkeld@chromium.org>