[autotest] Add a new field require_ssp in afe_jobs table
When a job is created, data in control file is parsed and saved in afe_jobs.
scheduler picks up the database row and tries to compile the commandline for
autoserv to run. require_ssp (require server-side packaging) is a new field
will be added to control file to force to use/do not use ssp when running
a server side test.

If the value is True or None(default), server-side packaging will be applied
to run server side test, if global config AUTOSERV/enable_ssp_container is
set to True.

Since the attribute can only be retrieved when parsing control file, it needs
to be stored in afe_jobs table. Therefore, the new field is added to the table.

require_ssp is nullable and has no default value, so the sql command should
only take several minutes to finish in production database.

BUG=chromium:453624
TEST=local run migrate.py sync and migrate.py sync 97

Change-Id: I509ae60f50c116df5ee76b2d302c35bb75551adc
Reviewed-on: https://chromium-review.googlesource.com/249170
Trybot-Ready: Dan Shi <dshi@chromium.org>
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Dan Shi <dshi@chromium.org>