Update reassociate time calculation
Previously, the reassociation time is determined based on the time when
scan is started to the time when client is connected. However, in the newer
version (2.3) of wpa_supplicant, the device will first try to reconnect
(disconnect then connect) without scanning.

Update the reassociate time calculation to consider both scan attempt
and disconnect attempt as the start of reassociation, whichever appears
first in the iw event log is considered the start of reassociation.

BUG=chromium:448155
TEST=Run network_WiFi_Reassociate test for both version of wpa_supplicant

Change-Id: I9c2c71999e16b5f1eab7d902450deb1221bf8f93
Reviewed-on: https://chromium-review.googlesource.com/250840
Trybot-Ready: Zeping Qiu <zqiu@chromium.org>
Tested-by: Zeping Qiu <zqiu@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Zeping Qiu <zqiu@chromium.org>