Fix time limit on reverify_unknown.
In reverify_unknown, the time limit for "idle too long" was set at
28 hours.  The arthmetic that figured that number was wrong.  The 28
hours figure was for "24 hours, plus a 4 hour buffer"; it should
have been figured as "24 hours, _less_ a _5_ hour buffer".  The
result was that DUTs that had been idle for for more than 20 hours
and less than 28 hours would sometimes be counted as broken.

BUG=chromium:460517
TEST=None

Change-Id: I8c4938790024e91e1fba74c141ee9c81ec7476e8
Reviewed-on: https://chromium-review.googlesource.com/251590
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>