REVERT LATER: autotest: Whitelist warning for power_Resume
This syscore_resume warning is something we need to get rid of,
but not something that is actively causing issues. In the
interest of unblocking results from this test, we'll whitelist
this particular warning for now. When the warning is fixed
we'll get rid of this hack.

BUG=chrome-os-partner:35778
TEST=run test, check for Whitelisted KernelError in logs

Change-Id: I68f571f91e2b39b4250c40805a13589d370a5404
Reviewed-on: https://chromium-review.googlesource.com/251832
Reviewed-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>