network_WiFi_ChannelScanDwellTime: disable scheduled scan before test
When scheduled scan is enabled, drivers (such as Intel WP2) that
support it could be in the middle of scheduled scan when we ask
the drvier to perform a single-channel scan. In this case the driver
may actually perform the single-channel scan after the scheduled
scan is completed. This causes the returned scan results for the
single-channel scan to contain BSSes from both scans, which would
mess up the dwell time calculation.

By default, scheduled scan is enabled. Disable it for this test
and renable it after the test is complete.

BUG=chromium:461069
TEST=Run this test
CQ-DEPEND=CL:252220

Change-Id: I5eccece60f723f49b4fd9de301d22b4e8e619297
Reviewed-on: https://chromium-review.googlesource.com/252342
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Zeping Qiu <zqiu@chromium.org>
Tested-by: Zeping Qiu <zqiu@chromium.org>