Use WebPageReplay in ChromeOS Video Tests.
WebPageReplay is an open source technology allowing users to prerecord web
traffic when navigating to a particular webpage. This traffic is recorded in
a .wpr file that can later be 'replayed' to serve that same content even when
no internet connection is available.

We are using it to pre-record a video and play it back during a test.

Pre-recording is not part of the test.

BUG=chromium:445636
TEST=Ran locally on a peach-pi.

Change-Id: I330648cb636a120529bfc8d85f43f3c08dcbc3e8
Reviewed-on: https://chromium-review.googlesource.com/255603
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Commit-Queue: Mussa Kiroga <mussa@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>