faft: Make sure the USB stick in servo matches the board under test
It's surprisingly easy to accidentally run FAFT with a wrong USB stick
plugged in. As long as it's any kind of test image, the pre-test checks
will pass... and when running tests that boot from USB, the fact that
the dev keys are a single keyset shared across all boards (unlike
(pre-)MP keys) means that ARM boards will happily try to boot x86
kernels and other nonsense like that, with results that look very much
like real test failures.

This patch adds a simple check to assert_test_image_in_usb_disk() to
make sure that the image on the USB stick is for the same board as the
image currently booted on the DUT (from local storage), which should
catch at least the most simple of mix-ups.

BUG=chrome-os-partner:37128
TEST=Ran firmware_RecoveryButton on a Jerry. Succeeded with a Jerry
image and failed with a Falco image plugged in.

Change-Id: Ifa5576afdec1bdd29ee872dcc8e987717138b97d
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/254432
Reviewed-by: Yusuf Mohsinally <mohsinally@chromium.org>