buffet: Check that buffet refreshes its access token if getting a 401.
We do this by requesting fake_device_server to mark the access token
used by buffet as invalid so requests from the device will error
out. This actually exposed a bug in buffet which is fixed in the
dependent CL.

Also rename fake_oauth.py to oauth.py to make room for a fake and use
that in unit-tests where needed.

BUG=brillo:275
TEST=The brillo_Registration autotest passes.
CQ-DEPEND=CL:255590

Change-Id: I0e7afeb9d1ff11fa0acbe52d88f7685021d5329c
Reviewed-on: https://chromium-review.googlesource.com/255480
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: David Zeuthen <zeuthen@chromium.org>
Tested-by: David Zeuthen <zeuthen@chromium.org>