Quit using `chvt` in video label detection.
The code for detecting what video labels to apply to a DUT used the
`chvt` command to work around a problem that could occur when a DUT
had been manually flipped to VT2.  The `chvt` command isn't
available on freon builds, and freon is the wave of the future.
Additionally, the `chvt` workaround isn't needed for freon anyway.

So, for all these reasons, kill the code.

BUG=chromium:464443
TEST=Run the code with atest host create and add_detected_labels.

Change-Id: I24ccf81995410bab9cea92b0452d5af52d39c50b
Reviewed-on: https://chromium-review.googlesource.com/256690
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>