network_WiFi_MissingBeacons: bring down AP interface to simulate AP disapperance
This is a much lightweight operation compare to the full AP teardown
(which takes more than 5 seconds to complete). This fix a possible race
where the wifi service is already removed from shill by the time we check
the state of that service. Also provides more accurate estimate for how
fast the client respond to beacon loss when AP disappears.

BUG=chromium:466355
TEST=Run this test

Change-Id: I056baae85c2ce6b0db25af56a3dc6780f94635d5
Reviewed-on: https://chromium-review.googlesource.com/259871
Reviewed-by: Zeping Qiu <zqiu@chromium.org>
Commit-Queue: Zeping Qiu <zqiu@chromium.org>
Tested-by: Zeping Qiu <zqiu@chromium.org>