autotest: Expand functionality of ProcessWatcher
Change ProcessWatcher to kill processes based on the name and arguments
to a process.  This allows us to run multiple copies of the binary with
slightly different arguments in parallel.

In some cases, failing to close output streams causes the
immediate death of the monitored process depending on what autotest
does next.  Fix this while we're here.

Sometimes we would like to start a process as a particular user,
(e.g. DBus system configs limit the set of users that can claim a
well known name).  Add MinijailConfig, which allows consumers to
start processes in a minijail.

BUG=brillo:535
TEST=n_faced_peerd starts up correctly inside a ProcessWatcher.

Change-Id: I30ae402425f3cb8174b51daba4e208dcdc6c0e79
Reviewed-on: https://chromium-review.googlesource.com/260588
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Alex Vakulenko <avakulenko@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>