build_RootFilesystemSize: Don't fail when using squashfs.
build_RootFilesystemSize relies on a test_image script to store the size
of the rootfs in production images in a file in /root in test images.
This approach doesn't work well with squashfs since the image is mounted
as read-write when this file is created, so the size is unknown.

This patch makes the test pass, reporting only the rootfs size on test
images when squashfs is used. This autotest test will be replaced by an
ImageTestStage test instead.

BUG=brillo:616
TEST=test_that build_RootFilesystemSize

Change-Id: I360fa09d9f20b8f1d601de6234257d7b5b3b3046
Reviewed-on: https://chromium-review.googlesource.com/261025
Reviewed-by: Bertrand Simonnet <bsimonnet@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Trybot-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>