Chameleon: Replug when CrOS fails to see any external display
We saw many time of the failure complaining CrOS failed to detect
the external display. Reruning the same test next time passed.
It is more like a flaky behavior and not a real issue. Replugging
the port is a common reaction for an user. So the test does the same.

BUG=chromium:443993
TEST=Ran display_ServerChameleonConnection and display_Resolution.

Change-Id: I5db8fd4ab57ebd3f634ce5da07116b018bfc614d
Reviewed-on: https://chromium-review.googlesource.com/261640
Trybot-Ready: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>