Chameleon: fix get_internal_resolution()
Currently, display_facade_native.get_internal_resolution() gets the
resolution of the internal screen from the system tool which may not
be the same as the one that's conveyed to the user. For example, on
samus, the raw resolution of the internal screen is 2560x1700 while
the user perceived one is 1280x850.

The problem when this occurs is that we will generate the test image
of the wrong size and then we will make wrong decision about whether
DUT outputs at the TV level. This can be reproduced by running
display_Resolution.mirrored on samus.

In this commit, we get the resolution from the display panel of Chrome
instead of relying on the system tool. It's the reliable way to get
the user perceived resolution of the screen.

BUG=chromium:469553
TEST=successfully ran display_Resolution.mirrored on samus, peach_pi
and peppy with related change 1d7651e.

Change-Id: If227476b493ea6412b3ea70eeb73c336fc7c0665
Reviewed-on: https://chromium-review.googlesource.com/261817
Tested-by: Hung-ying Tyan <tyanh@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>
Commit-Queue: Hung-ying Tyan <tyanh@chromium.org>