glbench: define alpha value in saved images
Alpha planes are saved and compared to reference images so we should
request them in the GL configuration, otherwise the results are
probably undefined.
Clear the alpha to 1 to match existing reference images.
Only the waffle code is updated because the egl and glx code is not
used with glbench and should be deleted soon anyway.

BUG=chromium:469861
TEST="test_that -b $BOARD $HOST graphics_SanAngeles" on link, daisy, lumpy

Change-Id: I9b4cc35c55a3506276f36e43f095a62a5fdd7c16
Reviewed-on: https://chromium-review.googlesource.com/261972
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Trybot-Ready: Ilja Friedel <ihf@chromium.org>
Tested-by: Ilja Friedel <ihf@chromium.org>
Commit-Queue: Frank Henigman <fjhenigman@chromium.org>