Chameleon: Don't cast to a tuple if no external display connected
The commit 031a087 changed the default return value of
get_external_resolution() from (0, 0) and (-1, -1) to None.
The adapter casts it to a tuple from a list (RPC only supports
a list). Should not cast it if it is a None.

BUG=chromium:469108
TEST=None. Trivial change.

Change-Id: I32ab3856219a2ac93ee871ca9f01513b37ad7344
Reviewed-on: https://chromium-review.googlesource.com/261968
Trybot-Ready: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>
Commit-Queue: Wai-Hong Tam <waihong@chromium.org>