Ignore network_WiFi_ChannelScanDwellTime test on panther router
Panther router doesn't support sending manually generated beacon
frames through monitor interface, which is required for this test.
So igonre this test if it is running against a panther router.

Also move the property getter for system board name to the base class
SiteLinuxSystem, which allows both the router and the client to
inherit it.

BUG=chromium:470194
TEST=Run network_WiFi_ChannelScanDwellTime test on both stumpy
     and panther routers

Change-Id: I201bc45ec4567b245517b0001a7444ecf81ad767
Reviewed-on: https://chromium-review.googlesource.com/262258
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Zeping Qiu <zqiu@chromium.org>
Tested-by: Zeping Qiu <zqiu@chromium.org>