[autotest] Create lab_inventory for inventory cron jobs
This converts the core of the run-board-inventory and
run-pool-inventory scripts to python.  The new scripts produce
output that is largely similar, but not identical to the old
bash scripts.

The scripts provide some performance advantage, and some flexibilty:
The bash scripts needed >75 minutes to generate board inventory; the
new python script can generate pool and board inventories at the
same time, and finishes in ~30 minutes.

BUG=None
TEST=unit tests, plus full command-line against cautotest

Change-Id: Id4f5443d7efb8b538bfe4453256201cce2f81bcf
Reviewed-on: https://chromium-review.googlesource.com/263594
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>