[autotest] Fix touch_UpdateErrors bug on link.
Prevent touch_UpdateErrors from catching a new kernel log which describes
the state of "gpu-standbox-failures-fatal" and "touch-calibration".
None of the errors I have seen that I'm interested in catching with this test
use the word "fail" without the word "error", so stop checking for it.

BUG=None
TEST=Ran it on link with and without the change.

Change-Id: I50dfada8ebc1b6ed7957b84c9355ebbc60a8cdf6
Reviewed-on: https://chromium-review.googlesource.com/262966
Reviewed-by: Scott Cunningham <scunningham@chromium.org>
Commit-Queue: Katherine Threlkeld <kathrelkeld@chromium.org>
Tested-by: Katherine Threlkeld <kathrelkeld@chromium.org>