Add code, fix bugs needed to run video frame image comparison tests using Chameleon.
This is a follow up to CL 262202 which laid the groundwork
for using chameleon in video frame image comparison tests for video.

Before these two changes, the control file control.chameleon.mp4.480p
would run a screenshot based image comparison test using Chameleon. After these two changes
this control file will actually use video frames from chameleon instead of
merely capturing a screenshot using chameleon.

Peach-pi was chosen because its 1080p definition results in mirror mode videos that
align well with image dimensions selected earlier.

So start with peach-pi, observe the results and make the test work with other devices.

To capture video with chameleon you need to play the video.
Add capability to play an HTML5 player from python.

Other fixes in this CL:
When capturing video frames with chameleon, we added code to
copy the frame image if it's the same as the previous one. However
the code did not added the path of the copied image to the list
of paths of images.

BUG=chromium:418217
TEST=Ran locally on a peach-pi.

Change-Id: I2c77aad51b6ee727e4aada1a675a2c235adb34db
Reviewed-on: https://chromium-review.googlesource.com/263800
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>
Commit-Queue: Mussa Kiroga <mussa@chromium.org>