power_MemorySuspend: allow 1% early wakeup
The goal of this test is to make sure that the memory content is unaltered
after suspend / resume. However this test failed a lot because of some
other component wake the machine up early.

This patch makes this test to not consider failing if we see less than 1%
early wakeup while running this test.

BUG=chromium:471911, chrome-os-partner:37417
TEST=manual. Test did not fail when early suspend is less than treshold.

Change-Id: I8b51e4ae5d605134a47fb79dd67ee3ae6a2316c7
Signed-off-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/264010
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>