Autotest: rewrite machine_install to be robust to devserver flakiness.
This CL both simplifies machine_install as well as adds protections
when devservers are flaky to provisioning.

1) I simplified machine_install by entirely removing the repair option
which was only referenced internally to this module as well as by one
autoupdate job that is not used anymore. I have remove that job in a previous
CL. Repair basically didn't need to exist. There was redundancy for it and
it's cleaner to just have a clean slate before ANY update.

2) Added a loop around the actual full update. Basically, I've added a #
of attempts for update. If an update fails, I will run a health check against
its devserver and re-attempt on another devserver if it's failed. This includes
re-staging the image on the other devserver as well as updating from it.

BUG=chromium:470581,chromium:471294
TEST=Moblab + Dynamic Suite + Instrumented Code to fail Autoupdate once to
check fallback + unittests in autoupdater.py + pyflakes/pylint on all
affected files.

Change-Id: I9836399269427e00fb6ccbcc79a08e0b11ef8859
Reviewed-on: https://chromium-review.googlesource.com/265954
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Chris Sosa <sosa@chromium.org>