Add a list of static APs to delay on DHCP.
There is a list of static APs in which the DHCP server takes a very
long time to come up even though the beacon is up.  This change
also moves the delay to be up after all of the pre-scans are complete.

What I noticed in testing is that some APs take a long time to even
beacon so delaying before that time really doesn't help.

TEST=Manual; Ran it
BUG=chrome-os-partner:37686

Change-Id: I51273c6fe0093f8b69859024f6d5367fb0c8c9ed
Reviewed-on: https://chromium-review.googlesource.com/265308
Trybot-Ready: Kris Rambish <krisr@chromium.org>
Tested-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Commit-Queue: Kris Rambish <krisr@chromium.org>