[autotest] Convert dut-status to get_host_special_tasks().
This is stage two of a three stage process to create a new
get_host_special_tasks() RPC call.  This change converts dut-status
to use the new API call, while maintaining compatibility with old
versions of the command.

This change can't be committed until the previous change is in prod.

BUG=None
TEST=Test dut-status on a local instance using the prod database

Change-Id: Icda500a83ffa1a98a9bf8d4b4fc5a0302d7673f6
Reviewed-on: https://chromium-review.googlesource.com/265574
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>