Send email notification for video related tests.
Send email notification to the address:
'chromeos-video-test-failures@google.com' for video related tests.

For those tests in the 'video' suite, we add BUG_TEMPLATE to the
suite with a 'cc' entry to set email forwarding.

For those tests not in the 'video' suite, we add a new 'cc' entry to
the BUG_TEMPLATE in the test's control file.

BUG=chromium:464666
TEST=None

Change-Id: I66275440febeffa1ea9def018a36a94629fe8d84
Reviewed-on: https://chromium-review.googlesource.com/265526
Reviewed-by: Owen Lin <owenlin@chromium.org>
Tested-by: Owen Lin <owenlin@chromium.org>
Commit-Queue: Owen Lin <owenlin@chromium.org>