[autotest] Create a new get_status_task RPC to improve dut_status.
Previously, status was determined by fetching all jobs in a given
time range, and then searching back through the returned list for
a job that indicated host status.  This had two drawbacks:
  * If the job indicating status was earlier than the start time
    of the query, no status could be determined.
  * The extra data returned in the query slowed it down.

This adds a new RPC that finds the task through a single database
query searching for the specific necessary condition.  This should
largely eliminate the '--' status, and is notably faster (possibly 4
times faster, or more).

This also changes the rules for what constitues a diagnostic job or
task, by changing the diagnosis associated with test jobs.  This
is done because
  * It makes determining status possible with a query to a single
    table.
  * The experience has been that test jobs don't indicate a
    diagnosis anyway.

This is the first of two parts; this change creates the server
side of the new RPC.  This change must be pushed to prod before
the follow-on change for the client side.

BUG=chromium:463632
TEST=Run dut_status using a local instance pointed at the prod database

Change-Id: Ifc61387f0cdc46ceb32062663c6a73c5d4bcfb4f
Reviewed-on: https://chromium-review.googlesource.com/265783
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>