autotest (wifi): unbuffer tcpdump output
By default, tcpdump buffers its output. This means that, if we kill
tcpdump, we lose frames that are still in its output buffer. This
is a problem for short tests (associate + get dhcp lease), as there
may not be enough traffic for tcpdump to flush out the association
frames.

Solve this by adding the '-U' option to the tcpdump command line.
This option tells tcpdump not to buffer output.

BUG=chromium:477536
TEST=manual test (see below)

Manual test
-----------
- ran network_WiFi_RegDomain
- examined /tmp/test_that_latest/results-1-network_WiFiRegdomain/network_WiFi_RegDomain.ca.hybrid/debug/router.0.pcap,
  verified that packet capture included association frames from client.

Change-Id: I8998e136342574c8d7cb58209cfcbf7d312b35d0
Reviewed-on: https://chromium-review.googlesource.com/266092
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>