Chameleon: raise Exception when an RPC call fails
Recently some RPC calls (e.g., set_mirrored(), load_calibration_image() etc)
failed with DevToolsClientUrlError but they are hidden in the return value
instead of a proper exception being raised. It makes it hard to triage the
bugs. This commit checks if a traceback string is returned from an RPC call
and raises an exception when it occurs.

BUG=chromium:478619,chromium:478870,chromium:478821,chromium:478793,chromium:478667 etc
TEST=ran display_Resolution and failed set_mirrored() deliberately

Change-Id: If713253f6085bab77b70d319507ef18328b38f0f
Reviewed-on: https://chromium-review.googlesource.com/266680
Tested-by: Hung-ying Tyan <tyanh@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Hung-ying Tyan <tyanh@google.com>