[autotest] Add a script to cleanup orphaned container.
If an autoserv process dies without being able to call handler of SIGTERM, the
container used to run the test will be orphaned. This adds overhead to the
drone. This script is used to clean up such containers.

This module also checks if the test job associated with a container has
finished. If so, kill the autoserv process for the test job and destroy the
container. To avoid racing condition, this only applies to job finished at least
1 hour ago.

BUG=chromium:479383
TEST=local test
1. create several server test, kill autoserv process with sudo kill -9
2. create several server test, manuall change database to mark the jobs are
   completed.
3. run the script, confirm all containers and autoserv processes are cleaned up.

Change-Id: Ib2cbba71ee40ba91582853a6cd3f91932b03c96d
Reviewed-on: https://chromium-review.googlesource.com/267052
Trybot-Ready: Dan Shi <dshi@chromium.org>
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Fang Deng <dengfang07@gmail.com>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Dan Shi <dshi@chromium.org>