gs_offloader: Only start offloading if gsutil has write access.
Adds a while-sleep loop to the initialization of gs_offloader
to ensure it can offload to the specified bucket. If a boto key
is installed after it has started, it will pass the check and
continue.

BUG=chromium:470565
TEST=launched moblab with no boto key, observed it spinning, then
installed a boto key and observed it offloading.

Change-Id: I2d3cf64ac588db1caec998bf33c9bff1a16e4cda
Reviewed-on: https://chromium-review.googlesource.com/267072
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>