Adding more logging to ui_SettingsPage to help investigate failure
We check if the DUT has a modem at the start of this test. If it does we
download the screenshot of Settings with a mobile connection. If it doesn't we
get the other one. It seems to be reporting no modem and then showing one.

BUG=chromium:482209
TEST=None

Change-Id: If5e2209274c22655e347d2f49b9136d3f9e6e659
Reviewed-on: https://chromium-review.googlesource.com/267747
Reviewed-by: Mussa Kiroga <mussa@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>