autotest: Fix up DHCP "Request Hostname" test
CL:266912 accidentally broke the rest of DHCP tests.  This was due
to comparing the ipconfig hostname to the supplied hostname parameter.
Unfortunately the undefined values for these two data were different:
for the ipconfig, this was an empty string ("") and for the supplied
configuration it was None.  This code provides a simple fix to only
validate the hostname parameter if it is actually provided in the
configuration.

While here, do a few more fix-ups to the network_DhcpRequestHostname
tests, notably to be more careful about resetting the hostname at the
end of the test.

BUG=chromium:482536
TEST=network_DhcpRequestHostName, network_DhcpNegotiationSuccess

Change-Id: I75f23710585cc1c30b9bb2c362c329b619998e93
Reviewed-on: https://chromium-review.googlesource.com/268313
Reviewed-by: Zeping Qiu <zqiu@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>