[autotest] Change adb_host check_host logic.
adb is about to be installed on all Chrome OS images. We want to
be able to determine if this device is a cros DUT or a host
testing an Android DUT. So now we require a magic file:
'/mnt/stateful_partition/.android_tester' to determine if we
want to use as an adb_host.

BUG=None
TEST=ran test_that locally and ensured the correct host type was
created.

Change-Id: I3948177b47e1a3c45b0bfe85525f2cd1fe057d76
Reviewed-on: https://chromium-review.googlesource.com/268355
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>