[autotest] Update servo based on the stable_version table.
This switches servo_host updates to use the afe_stable_versions
database table to select the target version for update.  This is
the same table used for selecting repair images for DUTs.

Previously, selection was done by asking the devserver for the
latest staged image for beaglebone_servo.  This would fail if the
devserver had not already staged the image.

Additionally, this removes some outstanding dead code that was
previously used to find the URL for a test image for use in servo
repair.

BUG=None
TEST=Force the flow on a local autotest instance; see servo update

Change-Id: I5cc4395c9b173e1dace13eee97f337b7f0a58b26
Reviewed-on: https://chromium-review.googlesource.com/268543
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Dan Shi <dshi@chromium.org>