[autotest] Retry container clone without using snapshot.
This CL attempts to alieviate the flakeness of lxc-clone in Ganeti instance.
We will evaluate the metadata to confirm the effectiveness, while looking for
the root cause of this issue.

Also add a property rootfs to container. This property will always point to the
folder of the container's rootfs. So the caller doesn't need to check if the
container is created from snapshot clone to determine if or not to use `delta0`
or `rootfs`.

BUG=chromium:484019
TEST=local test
sudo python site_utils/lxc_functional_test.py -v

Change-Id: Ibc04371bfb63e811a5bc14693d678adff6941323
Reviewed-on: https://chromium-review.googlesource.com/269009
Trybot-Ready: Dan Shi <dshi@chromium.org>
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Dan Shi <dshi@chromium.org>